+++ Box/DecodingTimeToSampleBox.h
@@ -13,11 +13,11 @@ public:
     std::vector<TimeToSampleItem> time_to_sample_item_table;
 
 public:
-    DecodingTimeToSampleBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    DecodingTimeToSampleBox(uint32_t size, uint32_t type, uint64_t start_pos);
 
     virtual ~DecodingTimeToSampleBox();
 
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
