+++ Box/DefaultContainerBox.h
@@ -9,10 +9,10 @@
 
 class DefaultContainerBox : public Box {
 public:
-    DefaultContainerBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    DefaultContainerBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~DefaultContainerBox();
 
-    virtual void Parse(struct Parser *parser, uint32_t start_pos) override;
+    virtual void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
