+++ Box/FreeSpaceBox.cpp
@@ -4,7 +4,7 @@
 
 #include "FreeSpaceBox.h"
 
-FreeSpaceBox::FreeSpaceBox(uint32_t size, uint32_t type, uint32_t start_pos) : Box(size, type, start_pos) {
+FreeSpaceBox::FreeSpaceBox(uint32_t size, uint32_t type, uint64_t start_pos) : Box(size, type, start_pos) {
     data = std::vector<uint8_t>(size - 8);
 }
 
@@ -12,6 +12,6 @@ FreeSpaceBox::~FreeSpaceBox() {
 
 }
 
-void FreeSpaceBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void FreeSpaceBox::Parse(struct Parser *parser, uint64_t start_pos) {
     for (size_t i = 0; i < size - 8; i++) data[i] = parser->file_reader->Read8();
 }
