+++ Box/MediaHeaderBox.h
@@ -19,11 +19,11 @@ public:
     uint16_t quality;
 
 public:
-    MediaHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    MediaHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos);
 
     virtual ~MediaHeaderBox();
 
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
