+++ Box/MetadataItemKeysBox.h
@@ -20,11 +20,11 @@ public:
     std::vector<KeyItem> key_item_table;
 
 public:
-    MetadataItemKeysBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    MetadataItemKeysBox(uint32_t size, uint32_t type, uint64_t start_pos);
 
     virtual ~MetadataItemKeysBox();
 
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
