+++ Box/SampleDescriptionBox.h
@@ -20,9 +20,9 @@ public:
     std::vector<SampleDescriptionItem> sample_description_item_table;
 
 public:
-    SampleDescriptionBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    SampleDescriptionBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~SampleDescriptionBox();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
