+++ Box/SyncSampleBox.h
@@ -12,9 +12,9 @@ class SyncSampleBox : public FullBox {
     std::vector<uint32_t> sync_sample_item_table;
 
 public:
-    SyncSampleBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    SyncSampleBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~SyncSampleBox();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
