+++ Box/TrackHeaderBox.cpp
@@ -4,7 +4,7 @@
 
 #include "TrackHeaderBox.h"
 
-TrackHeaderBox::TrackHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {
+TrackHeaderBox::TrackHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {
     matrix = std::vector<std::vector<uint32_t>>(3, std::vector<uint32_t>(3));
 }
 
@@ -12,7 +12,7 @@ TrackHeaderBox::~TrackHeaderBox() {
 
 }
 
-void TrackHeaderBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void TrackHeaderBox::Parse(struct Parser *parser, uint64_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     FileReader *file_reader = parser->file_reader;
